+++ src/main/java/com/umtdg/pfo/portfolio/fund/dto/PortfolioFundPred.java
@@ -1,6 +1,6 @@
 package com.umtdg.pfo.portfolio.fund.dto;
 
 public record PortfolioFundPred(
-    String code, String title, float price, int amount, float weight
+    String code, String title, double price, int amount, double weight
 ) {
 }
